Commit and Push....

+++ app/src/main/java/com/example/eventplannerapp/data/remote/EventApi.kt
@@ -7,6 +7,6 @@ interface EventApi {
 
     suspend fun addEvent(event: Event) : Event
 
-    suspend fun deleteEvent(eventId : Int) : Boolean
+    suspend fun deleteEvent(eventId : Long) : Boolean
 
 }
